+++ CHANGELOG.md
@@ -61,6 +61,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * New method signature for `PGS_Morphology.buffer()` that accepts a cap style parameter.
 * `PGS_PointSet.squareGrid()` now populates the grid with points upto and including `xMax` and `yMax` (inclusive).
 * Improved `PGS_Transformation.touchScale()` algorithm. It's now more performant and robust.
+* `PGS_Tiling.triangleSubdivision()` can now begin from both diagonal axes, not just the top-left to bottom-right diagonal.
 
 ### Fixed
 * `PGS_Morphology.rounding()` no longer gives invalid results.